requirements.txt
pydantic>=2.7.0
newsapi-python>=0.2.7
yfinance>=0.2.51
openai>=1.54.0
pandas>=2.2.0
python-dotenv>=1.0.1
tqdm>=4.66.4
requests>=2.32.0
pytest>=8.3.0
PyYAML>=6.0.2
fastapi>=0.116.0
uvicorn>=0.35.0
